CROSSVALIDATIO+MANUAL SPLITTING DATASET+PICKLES
I had splitted the dataset loading part from the classification part of the code, for a more practical use.
The loading of the dataset by the <code>**load_image_file**</code> need a bit of time, so saving it into the pickle just ones, let us speed the following classification part by accessing it.
The match_images  function is the one created by Davide and Vittoria, that build sequences of n Genuine signatures + 1 (G or F). But I had modified it to create 10-n G...GG and 10 G....GF sequences. It the code I had set n=2, so we are working on triples. The extention _V1 means that I'm am using the former function. Davide Gurnari had developed another one, but I didn't tested it yet. It is denoted by _V2.  
The crossvalidation is implemented thanks the GridSearchCV function, focusing on the accuracy parameter. (It automatically balances the Dataset thanks to the attibute weight='balance' provided by Davide Gurnari) It takes a very long time on the SVM classifier and RF Classifier. However it does not take long time for the LR classifier. 
You can see the report results, the best parameters found by the crossvalidation and its time of execution on each classifier. in the file results.txt. The only one missing is the HOG+SVM because it takes to much time (like 5 h).